+++ .kotlin-template/src/main/kotlin/com/theagilemonkeys/ellmental/changeme/Main.kt
@@ -1,0 +1,5 @@
+package com.theagilemonkeys.ellmental.changeme
+
+fun main() {
+    println("Hello World!")
+}
