--- .kotlin-template/src/main/kotlin/com/theagilemonkeys/llmental/changeme/Main.kt
@@ -1,5 +1,0 @@
-package com.theagilemonkeys.llmental.changeme
-
-fun main() {
-    println("Hello World!")
-}
