+++ .kotlin-template/src/test/kotlin/com/theagilemonkeys/ellmental/changeme/MainKtTest.kt
@@ -1,0 +1,12 @@
+package com.theagilemonkeys.ellmental.changeme
+
+import io.kotest.core.spec.style.StringSpec
+import io.kotest.property.forAll
+
+class MainKtTest : StringSpec({
+    "String size" {
+        forAll<String, String> { a, b ->
+            (a + b).length == a.length + b.length
+        }
+    }
+})
