+++ core/src/main/kotlin/com/theagilemonkeys/ellmental/core/schema/Embedding.kt
@@ -1,0 +1,7 @@
+package com.theagilemonkeys.ellmental.core.schema
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+@JvmInline
+value class Embedding(val value: List<Double>)
