+++ examples/semanticsearchservicedemo/.env.template
