--- llmental-core/src/main/kotlin/com/theagilemonkeys/llmental/core/schema/Embedding.kt
@@ -1,7 +1,0 @@
-package com.theagilemonkeys.llmental.core.schema
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-@JvmInline
-value class Embedding(val value: List<Double>)
