--- llmental-core/src/main/kotlin/com/theagilemonkeys/llmental/core/schema/Id.kt
@@ -1,7 +1,0 @@
-package com.theagilemonkeys.llmental.core.schema
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-@JvmInline
-value class Id(val value: String)
