--- llmental-core/src/main/kotlin/com/theagilemonkeys/llmental/core/schema/SemanticEntry.kt
@@ -1,17 +1,0 @@
-package com.theagilemonkeys.llmental.core.schema
-
-import kotlinx.serialization.Serializable
-import java.util.*
-
-@Serializable
-data class SemanticEntry(
-    val id: Id = generateId(),
-    val embedding: Embedding,
-    val metadata: Map<String, String>? = null,
-    val clusterId: Id? = null
-) {
-    companion object {
-        fun generateId() = Id(UUID.randomUUID().toString())
-    }
-}
-
