--- modules/embeddings-model/src/main/kotlin/com/theagilemonkeys/llmental/embeddingsModel/EmbeddingsModel.kt
@@ -1,7 +1,0 @@
-package com.theagilemonkeys.llmental.embeddingsModel
-
-import com.theagilemonkeys.llmental.core.schema.Embedding
-
-interface EmbeddingsModel<Params> {
-    suspend fun embed(text: String, params: Params? = null): Embedding
-}
