--- modules/embeddings-model/src/test/kotlin/com/theagilemonkeys/llmental/semanticSearchServiceDemo/MainKtTest.kt
@@ -1,12 +1,0 @@
-package com.theagilemonkeys.llmental.semanticSearchServiceDemo
-
-import io.kotest.core.spec.style.StringSpec
-import io.kotest.property.forAll
-
-class MainKtTest : StringSpec({
-    "String size" {
-        forAll<String, String> { a, b ->
-            (a + b).length == a.length + b.length
-        }
-    }
-})
