+++ modules/embeddingsmodel/src/main/kotlin/com/theagilemonkeys/ellmental/embeddingsmodel/EmbeddingsModel.kt
@@ -1,0 +1,7 @@
+package com.theagilemonkeys.ellmental.embeddingsmodel
+
+import com.theagilemonkeys.ellmental.core.schema.Embedding
+
+interface EmbeddingsModel<Params> {
+    suspend fun embed(text: String, params: Params? = null): Embedding
+}
