--- modules/vector-store/src/main/kotlin/com/theagilemonkeys/llmental/vectorStore/VectorStore.kt
@@ -1,13 +1,0 @@
-package com.theagilemonkeys.llmental.vectorStore
-
-import com.theagilemonkeys.llmental.core.schema.SemanticEntry
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class QueryOutput(val entries: List<SemanticEntry>)
-
-interface VectorStore {
-    suspend fun upsert(semanticEntry: SemanticEntry)
-
-    suspend fun query(semanticEntry: SemanticEntry): QueryOutput
-}
