--- semantic-search-service-demo/build.gradle.kts
@@ -1,21 +1,0 @@
-group = "com.theagilemonkeys.llmental.semantic-search-service-demo"
-
-val http4kVersion: String by project
-
-plugins {
-    application
-    kotlin("plugin.serialization") version "1.8.22"
-}
-
-application {
-    mainClass.set("com.theagilemonkeys.semanticSearchServiceDemo.MainKt")
-}
-
-dependencies {
-    implementation(projects.llmentalCore)
-    implementation(projects.modules.embeddingsModel)
-    implementation(projects.modules.vectorStore)
-    implementation(projects.modules.semanticSearch)
-    implementation(platform("org.http4k:http4k-bom:$http4kVersion"))
-    implementation("org.http4k:http4k-core")
-}
