+++ settings.gradle.kts
@@ -1,10 +1,10 @@
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 
-rootProject.name = "llmental"
+rootProject.name = "ellmental"
 include(
-    "llmental-core",
-    "modules:embeddings-model",
-    "modules:vector-store",
-    "modules:semantic-search",
-    "semantic-search-service-demo"
+    "core",
+    "modules:embeddingsmodel",
+    "modules:vectorstore",
+    "modules:semanticsearch",
+    "examples:semanticsearchservicedemo"
 )
